splitted base system and wrapper
