+++ installation/install_GUI_scripts.sh
